Use Typescript in Cypress tests
