--- tests/ui/tests/config.js
@@ -1,5 +1,0 @@
-const defaultKymaDashboardAddress = "http://localhost:3001";
-
-export default {
-  clusterAddress: Cypress.env("DOMAIN") || defaultKymaDashboardAddress
-};
