+++ tests/ui/tests/config.ts
@@ -1,0 +1,6 @@
+const defaultKymaDashboardAddress = "http://localhost:3001";
+
+
+export default {
+  clusterAddress: (Cypress.env("DOMAIN") as string) || defaultKymaDashboardAddress
+};
