+++ tests/ui/tests/package.json
@@ -9,8 +9,11 @@
     "start-k3d": "./scripts/k3d-local-dev.sh"
   },
   "devDependencies": {
-    "cypress-file-upload": "^5.0.8",
-    "js-yaml": "^4.1.0"
+    "@kubernetes/client-node": "0.20.0",
+    "cypress-file-upload": "5.0.8",
+    "js-yaml": "4.1.0",
+    "js-yaml-js-types": "1.0.1",
+    "typescript": "5.3.3"
   },
   "optionalDependencies": {
     "cypress": "^12.3.0"
