--- tests/ui/tests/support/filter.js
@@ -1,3 +1,0 @@
-Cypress.Commands.add('filterWithNoValue', { prevSubject: true }, $elements =>
-    $elements.filter((_, e) => !e.value),
-);
