--- tests/ui/tests/support/index.js
@@ -1,10 +1,0 @@
-import './exceptions';
-import './login';
-import './navigation';
-import './loadFile';
-import './views';
-import './random';
-import './filter';
-import './list';
-import './buttons';
-
