+++ tests/ui/tests/support/views/namespace.ts
