+++ tests/ui/tests/tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "compilerOptions": {
+    "target": "es5",
+    "lib": ["es5", "dom"],
+    "types": ["cypress", "node"],
+    // We will start with strict false for now. After we cleaned up the code, we should set it to true.
+    "strict": false,
+  },
+  "include": ["./**/*.ts", "./**/*.spec.ts"]
+}
